web/st/back-end/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-stubs==1.1.0
gunicorn==20.0.4
idna==2.8
Markdown==3.1.1
pytz==2019.3
sqlparse==0.3.0
urllib3==1.25.8
psycopg2==2.7.3.1
docutils==0.16